Move `getCommandMap` to new class (`utils/ServerCommandMap.kt`)
